fix: artifact filtering and DVC history sync (refactored)
- Centralized Git/DVC logic in `src/scheduler/utils/git_dvc.py`.
- Improved network resilience with 60s timeout for `git fetch`.
- Enhanced error handling: routes now return descriptive errors (503/500) for sync failures instead of empty lists.
- Implemented DVC-aware artifact filtering by parsing `dvc.yaml`.
- Fixed "ghost" stages in DVC history by resolving revisions to SHAs and adding `git clean -fd` in the worker agent.
- Updated `test_portal.py` and `pyproject.toml`.
- Synchronized `remote_headnode_service.py` and `headnode_service.py`.
