+++ artifacts/random_table.csv
@@ -1,11 +1,11 @@
 epoch,loss,accuracy
-1,0.3031,0.8463
-2,0.0583,0.7755
-3,0.4427,0.8353
-4,0.2451,0.8119
-5,0.2152,0.7125
-6,0.0848,0.8621
-7,0.4807,0.7143
-8,0.0105,0.7639
-9,0.4006,0.9617
-10,0.1398,0.7816
+1,0.4527,0.8424
+2,0.154,0.9866
+3,0.1592,0.8734
+4,0.4539,0.7257
+5,0.1471,0.936
+6,0.1526,0.9011
+7,0.4263,0.9026
+8,0.2055,0.9514
+9,0.0164,0.8938
+10,0.3694,0.961
