+++ dvc.lock
@@ -41,9 +41,9 @@ stages:
       size: 70
     - path: artifacts/random_plot.png
       hash: md5
-      md5: 679c3407f6f42fdc3831d521f4231b45
+      md5: b7845e6f8f327b28ff30e30e23155344
       size: 86
     - path: artifacts/random_table.csv
       hash: md5
-      md5: 3d28f90ddb3be3fdb097e3c5da216268
-      size: 192
+      md5: e7cbca65c20de8d13097f5df414482a6
+      size: 189
