+++ src/scheduler/headnode_service.py
@@ -18,7 +18,6 @@ import shutil
 import requests
 import subprocess
 import sys
-import yaml
 import time
 import threading
 import re
@@ -27,6 +26,7 @@ import tempfile
 from urllib.parse import urlparse
 from dotenv import load_dotenv
 from werkzeug.middleware.proxy_fix import ProxyFix
+from utils.git_dvc import find_local_repo, resolve_revision, get_dvc_artifacts, filter_artifact_files, REPOS_DIR, GitError
 
 load_dotenv()
 
@@ -511,25 +511,6 @@ def check_space():
         "sufficient": free_gb > FREE_SPACE_THRESHOLD_GB
     })
 
-REPOS_DIR = os.path.join(os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__)))), "repositories")
-
-def find_local_repo(repo_slug):
-    """Find the local clone of a repo, handling owner name mismatches."""
-    # Try exact match first
-    exact = os.path.join(REPOS_DIR, repo_slug)
-    if os.path.exists(exact) and os.path.exists(os.path.join(exact, ".git")):
-        return exact
-    
-    # Fallback: search by repo name only across all owner dirs
-    repo_name = repo_slug.split('/')[-1] if '/' in repo_slug else repo_slug
-    if os.path.exists(REPOS_DIR):
-        for owner_dir in os.listdir(REPOS_DIR):
-            if owner_dir.startswith('_'):  # Skip _tmp_artifacts etc.
-                continue
-            candidate = os.path.join(REPOS_DIR, owner_dir, repo_name)
-            if os.path.isdir(candidate) and os.path.exists(os.path.join(candidate, ".git")):
-                return candidate
-    return None
 
 @app.route('/artifacts/<repo_owner>/<repo_name>/<rev>/<path:file_path>', methods=['GET'])
 def artifacts(repo_owner, repo_name, rev, file_path):
@@ -1121,16 +1102,11 @@ h1 {{ font-size: 1.25rem; margin: 0; color: #38bdf8; }}
 
     # Resolve rev to an absolute commit SHA to prevent "zombie" stages when a branch moves
     local_repo_path = find_local_repo(repo_full_name)
-    if local_repo_path:
-        try:
-            # First, fetch to make sure we have the latest SHAs for branches
-            subprocess.run(["git", "fetch", "--all"], cwd=local_repo_path, capture_output=True, timeout=5)
-            # Resolve the revision
-            res = subprocess.run(["git", "rev-parse", rev], cwd=local_repo_path, capture_output=True, text=True, timeout=5)
-            if res.returncode == 0:
-                rev = res.stdout.strip()
-        except Exception as e:
-            app.logger.warning(f"Could not resolve revision {rev} to SHA for {repo_full_name}: {e}")
+    try:
+        rev = resolve_revision(local_repo_path, rev)
+    except GitError as e:
+        app.logger.warning(f"Revision resolution failed for {repo_full_name} ({rev}): {e}")
+        # Fallback to symbolic name if resolution fails, but we already tried fetching
 
     with remote_viewers_lock:
         if repo_full_name in remote_viewers:
@@ -1293,79 +1269,25 @@ def api_latest_artifacts(repo):
         result = subprocess.run(cmd, capture_output=True, text=True, cwd=local_repo_path, timeout=5)
         
         if result.returncode != 0:
-            # If the commit is not locally known, try fetching it quickly (timeout 5s)
-            app.logger.info(f"Commit not found locally, fetching for {repo}...")
-            subprocess.run(["git", "fetch", "--all"], cwd=local_repo_path, capture_output=True, timeout=5)
-            result = subprocess.run(cmd, capture_output=True, text=True, cwd=local_repo_path, timeout=5)
+            # If the commit is not locally known, try resolving it (includes fetch)
+            try:
+                commit_hash = resolve_revision(local_repo_path, commit_hash)
+                result = subprocess.run(cmd, capture_output=True, text=True, cwd=local_repo_path, timeout=5)
+            except GitError as ge:
+                return jsonify({"error": f"Repository sync failed: {str(ge)}"}), 503
 
         if result.returncode == 0:
             all_git_files = [l for l in result.stdout.strip().split("\n") if l]
-
-            # Attempt to discover DVC artifacts from dvc.yaml
-            dvc_artifacts = set()
-            try:
-                show_cmd = ["git", "show", f"{commit_hash}:dvc.yaml"]
-                show_res = subprocess.run(show_cmd, capture_output=True, text=True, cwd=local_repo_path, timeout=5)
-                if show_res.returncode == 0:
-                    dvc_config = yaml.safe_load(show_res.stdout)
-                    if dvc_config and 'stages' in dvc_config:
-                        for stage_name, stage_cfg in dvc_config['stages'].items():
-                            for key in ['outs', 'metrics', 'plots']:
-                                if key in stage_cfg:
-                                    items = stage_cfg[key]
-                                    if isinstance(items, list):
-                                        for item in items:
-                                            if isinstance(item, str):
-                                                dvc_artifacts.add(item)
-                                            elif isinstance(item, dict):
-                                                # Handle format: {"path": {"cache": false}}
-                                                for path in item.keys():
-                                                    dvc_artifacts.add(path)
-                elif show_res.returncode != 0:
-                    app.logger.info(f"dvc.yaml not found in {repo} at {commit_hash}")
-            except Exception as e:
-                app.logger.warning(f"Failed to parse dvc.yaml for {repo}: {e}")
-
-            files = []
-            for line in all_git_files:
-                # 1. If we have DVC artifacts, strictly filter by them
-                if dvc_artifacts:
-                    # Check if the file is one of the artifacts or inside an artifact directory
-                    is_artifact = False
-                    for art in dvc_artifacts:
-                        if line == art or line.startswith(art + '/'):
-                            is_artifact = True
-                            break
-                    if not is_artifact:
-                        continue
-                else:
-                    # 2. Fallback: Aggressive exclusion of non-artifact files
-                    # Filter out system and config directories
-                    if any(line.startswith(p) for p in [".git/", ".github/", ".dvc/", ".idea/", ".vscode/", ".agent/"]):
-                        continue
-                    # Filter out common source/config extensions
-                    if any(line.endswith(ext) for ext in [
-                        ".py", ".sh", ".md", ".yaml", ".yml", ".json",
-                        ".toml", ".lock", ".txt", ".gitattributes", ".gitignore", ".dvcignore"
-                    ]):
-                        # Exception for common metric/plot suffixes even in fallback
-                        if not any(x in line.lower() for x in ["metric", "plot", "result", "output", "artifact"]):
-                            continue
-
-                files.append({
-                    "path": line,
-                    "is_dir": False,
-                    "size": 0,
-                    "isout": True
-                })
+            dvc_artifacts = get_dvc_artifacts(local_repo_path, commit_hash)
+            files = filter_artifact_files(all_git_files, dvc_artifacts)
             return jsonify(files)
         else:
             app.logger.error(f"Git ls-tree failed for {repo}: {result.stderr}")
-            return jsonify([])
+            return jsonify({"error": f"Failed to list repository files: {result.stderr}"}), 500
 
     except Exception as e:
         app.logger.error(f"Error listing latest artifacts from local Git for {repo}: {e}")
-        return jsonify([])
+        return jsonify({"error": f"Internal error while listing artifacts: {str(e)}"}), 500
 
 @app.route('/api/projects/<path:repo>/artifact/history', methods=['GET'])
 def api_artifact_history(repo):
